SecureLogger/requirements.txt
pynput
numpy
pandas
scikit-learn
joblib
pycryptodome